backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
dj-database-url
psycopg2-binary
whitenoise
python-dotenv
gunicorn
cloudinary==1.44.2
django-cloudinary-storage==0.3.0
